+++ Extensions.Core/VersionHistory.md
@@ -1,4 +1,7 @@
 ﻿# Extensions.Core Version History.
 
-### **4.9.700 (2023-03-29)**<br>
-	- Separate `System.Core` class for logging functionality from Extensions.cs.
+### **4.9.700 (2023-03-28)**<br>
+	- Separate `Extensions.Identity` class from Extensions.cs.
+
+### **6.0.800 (2024-02-22)**<br>
+	- Rearchitected for Extensions scaffolding design.